% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAGxEPlus_functions_MYZ_2024-09-19.R
\name{SPAGxE_Plus_Nullmodel}
\alias{SPAGxE_Plus_Nullmodel}
\title{Fits a genotype-independent model of SPAGxE+}
\usage{
SPAGxE_Plus_Nullmodel(
  traits = "survival/binary/quantitative",
  formula = NULL,
  data = NULL,
  pIDs = NULL,
  gIDs = NULL,
  range = c(-100, 100),
  length.out = 10000,
  sparseGRM,
  E,
  ...
)
}
\arguments{
\item{traits}{a character value corresponding to phenotype. It should be "binary" for binary phenotype, "survival" for time-to-event phenotype, and "others" for other types of phenotype.}

\item{formula}{a formula to be passed to function lm(), glm(), or coxph(). For more details, please refer to package survival.}

\item{data}{a data.frame in which to interpret the variables named in the formula}

\item{pIDs}{a character vector of subject IDs. NOTE: its order should be the same as the subjects order in the formula.}

\item{gIDs}{a character vector of subject IDs. NOTE: its order should be the same as the subjects order of the Geno.mtx (i.e. the input of the function SPAGxE_CCT()).}

\item{range}{a two-element numeric vector (default: c(-100,100)) to specify the domain of the empirical CGF.}

\item{length.out}{a positive integer (default: 9999) for empirical CGF. Larger length.out corresponds to longer calculation time and more accurate estimated empirical CGF.}

\item{sparseGRM}{a three-column sparse GRM file with the first column as "ID1",the second column as "ID2", and the last column as "Value" (i.e., two times of kinship coefficient) without information of distant genetic relatedness (such as population structure).}

\item{E}{a numeric environmental factor with each element as an environmental factor value of an individual.}

\item{...}{Other arguments passed to function glm() or coxph(). For more details, please refer to package survival.}
}
\value{
Residuals after fitting a genotype-independent (covariate-only) model.
}
\description{
Fits a null linear regression model for quantitative traits, a null logistic regression model for binary traits ,or a null Cox proportional hazards model for time-to-event traits and then calculates residuals under a genotype-independent model.
}
