% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAGxEPlus_functions_MYZ_2024-12-03.R
\name{SPAGxE_Plus_one_SNP}
\alias{SPAGxE_Plus_one_SNP}
\title{A scalable and accurate framework to account for familial relatedness for large-scale genome-wide gene-environment interaction (GxE) analysis (One-SNP-version).}
\usage{
SPAGxE_Plus_one_SNP(
  g,
  E,
  Phen.mtx,
  Cova.mtx,
  obj.SPAGxE_Plus_Nullmodel,
  epsilon = 0.001,
  Cutoff = 2,
  impute.method = "fixed",
  missing.cutoff = 0.15,
  min.maf = 1e-05,
  G.model = "Add"
)
}
\arguments{
\item{g}{a numeric genotype vector. Missing genotype should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{others}{the same as function SPAGxE_Plus(). NOTE that we do not check subject order in this one-snp-version !!!}
}
\value{
the same as function SPAGxE_Plus().
}
\description{
One-SNP-version SPAGxE_Plus() function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
