% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmixGxECCTPlus_functions_MYZ_2024-09-17.R
\name{SPAGxEmix_CCT_Plus}
\alias{SPAGxEmix_CCT_Plus}
\title{A scalable and accurate analytical framework to account for both population structure and family relatedness in large-scale genome-wide gene-environment interaction (GxE) analysis in admixed populations.}
\usage{
SPAGxEmix_CCT_Plus(
  traits = "binary/quantitative",
  Geno.mtx,
  ResidMat,
  E,
  Phen.mtx,
  Cova.mtx,
  topPCs,
  sparseGRM,
  epsilon = 0.001,
  Cutoff = 2,
  impute.method = "fixed",
  missing.cutoff = 0.15,
  min.maf = 0.001,
  G.model = "Add",
  topPCs.pvalue.cutoff = 0.05,
  MAF.est.negative.ratio.cutoff = 0.1
)
}
\arguments{
\item{Geno.mtx}{a numeric genotype matrix with each row as an individual and each column as a genetic variant.
Column names of genetic variations and row names of subject IDs are required.
Missing genotypes should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{ResidMat}{a two-column data frame with the first column as "SubjID" and the second column as model residuals after fitting a genotype-independent model (i.e., a covariate-only model in which marginal genetic effect and GxE effect are 0)}

\item{E}{a numeric environmental factor with each element as an environmental factor value of an individual.}

\item{Phen.mtx}{phenotype dataframe at least including three columns of ID, surv.time and event for time-to-event trait analysis, two columns of ID and linear phenotype Y for linear trait analysis, two columns of ID and binary phenotype Y for binary trait analysis, or two columns of ID and ordinal categorical phenotype Y for ordinal categorical trait analysis.}

\item{Cova.mtx}{a covariate matrix excluding the environmental factor E.}

\item{topPCs}{a covariate matrix including the SNP-derived principle components (PCs) containing all information of population structure (without recent genetic relatedness or kinship information).}

\item{sparseGRM}{a three-column sparse GRM file with the first column as "ID1",the second column as "ID2", and the last column as "Value" (i.e., two times of kinship coefficient) without information of distant genetic relatedness (such as population structure).}

\item{epsilon}{a numeric value (default: 0.001) to specify the p-value cutoff for betaG estimation. Please see details for more information.}

\item{Cutoff}{a numeric value (Default: 2) to specify the standard deviation cutoff to be used.
If the test statistic lies within the standard deviation cutoff, its p value is calculated based on a normal distribution approximation,
otherwise, its p value is calculated based on a saddlepoint approximation.}

\item{impute.method}{a character string (default: "fixed") to specify the method to impute missing genotypes.
"fixed" imputes missing genotypes (NA) by assigning the mean genotype value.}

\item{missing.cutoff}{a numeric value (default: 0.15) to specify the cutoff of the missing rates.
Any variant with missing rate higher than this cutoff will be excluded from the analysis.}

\item{min.maf}{a numeric value (default: 0.001) to specify the cutoff of the minimal MAF. Any SNP with MAF < cutoff will be excluded from the analysis.}

\item{G.model}{model type}
}
\value{
an R matrix with the following columns
\item{MAF}{Minor allele frequencies calculated with half of mean value of genotypes}
\item{missing.rate}{Missing rates}
\item{p.value.spaGxEmix.plus}{p value from SPAGxE method}
\item{p.value.spaGxEmix.plus.Wald}{p value from SPAGxE_Wald method}
\item{p.value.spaGxEmix.plus.CCT.Wald}{p value (recommanded) from SPAGxE_CCT method}
\item{p.value.normGxEmix.plus}{p value from NormGxE method (based on a normal distribution approximation)}
\item{p.value.betaG}{p value of the marginal genetic effect based on a normal distribution approximation}
\item{Stat.betaG}{score statistics testing for marginal genetic effect}
\item{Var.betaG}{estimated variances of the score statistics testing for marginal genetic effect}
\item{z.betaG}{z values (using Var1) corresponding to the score statistics testing for marginal genetic effect}
\item{MAF.est.negative.num}{numbers of negative individual-level MAF estimated using linear regression model}
\item{MAC}{minor allele counts}
}
\description{
A scalable and accurate analytical framework to account for both population structure and family relatedness in large-scale genome-wide gene-environmental interaction (GxE) analyses of quantitative traits and binary traits.
}
\details{
To run SPAGxEmix_CCT_Plus, the following two steps are required:
\itemize{
  \item Step 1: Use function SPA_G_Null_Model() or other functions to fit a genotype-independent (covariate-only) model to get residuals under a genotype-independent (covariate-only) model.
  \item Step 2: Use function SPAGxEmix_CCT_Plus() to calculate p value for each genetic variant to conduct a GxE analysis.
}

SPAGxEmix_CCT_Plus() is an extension of SPAGxEmix_CCT() which additionally uses sparse GRM to account for family relatedness in an admixed population or multiple populations.
SPAGxEmix_CCT_Plus() uses a hybrid strategy with both saddlepoint approximation and normal distribution approximation.
Generally speaking, saddlepoint approximation is more accurate than, but a little slower than, the traditional normal distribution approximation.
Hence, when the score statistic is close to 0 (i.e. p-values are not small), we use the normal distribution approximation.
And when the score statistic is far away from 0 (i.e. p-values are small), we use the saddlepoint approximation.
Argument 'Cutoff' is to specify the standard deviation cutoff.

Sometimes, the order of subjects between phenotype data and genotype data are different, which could lead to some errors.
To avoid that, we ask users to specify the IDs of both phenotype data (pIDs) and genotype data (gIDs) when fitting the null model.
Users are responsible to check the consistency between pIDs and formula, and the consistency between gIDs and Geno.mtx.
}
