% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmixGxECCTPlus_functions_MYZ_2024-09-17.R
\name{SPAGxEmix_CCT_Plus_one_SNP}
\alias{SPAGxEmix_CCT_Plus_one_SNP}
\title{A scalable and accurate framework to account for both population structure and family relatedness for large-scale genome-wide gene-environment interaction (GxE) analysis in admixed populations (One-SNP-version).}
\usage{
SPAGxEmix_CCT_Plus_one_SNP(
  traits = "binary/quantitative",
  g,
  ResidMat,
  E,
  Phen.mtx,
  Cova.mtx,
  topPCs,
  sparseGRM,
  epsilon = 0.001,
  Cutoff = 2,
  impute.method = "fixed",
  missing.cutoff = 0.15,
  min.maf = 0.001,
  G.model = "Add",
  topPCs.pvalue.cutoff = 0.05,
  MAF.est.negative.ratio.cutoff = 0.1
)
}
\arguments{
\item{g}{a numeric genotype vector. Missing genotype should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{others}{the same as function SPAGxEmix_CCT_Plus(). NOTE that we do not check subject order in this one-snp-version !!!}
}
\value{
the same as function SPAGxEmix_CCT_Plus().
}
\description{
One-SNP-version SPAGxEmix_CCT_Plus() function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
