% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAGxECCT.R
\name{SPAGxEmixCCT_localance_one_SNP}
\alias{SPAGxEmixCCT_localance_one_SNP}
\title{A scalable and accurate analysis framework for a large-scale genome-wide gene-environmental interaction (GxE) analyses of quantitative traits, binary traits, time-to-event traits, and ordinal categorical traits to identify ancestry-specific GxE effects by incorporating local ancestry (One-SNP-version).}
\usage{
SPAGxEmixCCT_localance_one_SNP(
  traits = "survival/binary/quantitative/categorical",
  g,
  haplo_numVec,
  R,
  E,
  Phen.mtx,
  Cova.mtx,
  Cova.haplo.mtx,
  epsilon = 0.001,
  Cutoff = 2,
  impute.method = "fixed",
  missing.cutoff = 0.15,
  min.maf = 1e-05,
  G.model = "Add"
)
}
\arguments{
\item{g}{a numeric ancestry-specific genotype vector. Missing genotype should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{others}{the same as function SPAGxEmixCCT_localance(). NOTE that we do not check subject order in this one-snp-version !!!}
}
\value{
the same as function SPAGxEmixCCT_localance().
}
\description{
One-SNP-version SPAGxEmixCCT_localance() function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
