% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAGxECCT.R
\name{SPAGxE_CCT_one_SNP}
\alias{SPAGxE_CCT_one_SNP}
\title{SaddlePoint Approximation implementation of gene-environmental interaction (GxE) analysis of complex traits (One-SNP-version).}
\usage{
SPAGxE_CCT_one_SNP(
  traits = "survival/binary/quantitative/categorical",
  g,
  R,
  E,
  Phen.mtx,
  Cova.mtx,
  epsilon = 0.001,
  Cutoff = 2,
  impute.method = "fixed",
  missing.cutoff = 0.15,
  min.maf = 1e-05,
  G.model = "Add"
)
}
\arguments{
\item{g}{a numeric genotype vector. Missing genotype should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{others}{the same as function SPAGxE_CCT_surv. NOTE that we do not check subject order in this one-snp-version !!!}
}
\value{
the same as function SPAGxE_CCT.
}
\description{
One-SNP-version SPAGxE_CCT function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
